+++ packages/server/src/auth/user.js
@@ -7,15 +7,19 @@ export const matchesUser = (userMap, username, password) => {
   return userMap[username].testPassword(password)
 }
 
-const sha1Base64 = text => {
-  var digest = crypto.createHash('sha1');
+const hashBase64 = (text, hash = 'sha1') => {
+  var digest = crypto.createHash(hash);
   digest.update(text, 'utf8');
   return digest.digest('base64');
 }
 
 const getTestPassword = password => {
-  if (password.startsWith('{SHA}')) {
-    return pwd => sha1Base64(pwd) == password.slice(5)
+  if (password.startsWith('{SHA256-salted}')) {
+    const pos = password.indexOf('.')
+    const salt = Buffer.from(password.substring(15, pos), 'base64').toString()
+    return pwd => hashBase64(salt + pwd, 'sha256') == password.substring(pos + 1)
+  } else if (password.startsWith('{SHA}')) {
+    return pwd => hashBase64(pwd) == password.slice(5)
   } else if (password.startsWith('{PLAIN}')) {
       return pwd => pwd == password.slice(7)
   } else {