+++ packages/server/src/auth/user.test.js
@@ -22,6 +22,13 @@ t.test('sha1 formatted password', t => {
   t.end()
 })
 
+t.test('sha256salted formatted password', t => {
+  const userMap = users2UserMap([{username: 'admin', password: '{SHA256-salted}lxBgzWvrFbDD+pcA.wyBh/87lu8SXrZ7Af0oovObiZeJk14AoBTxDW7HT5pY='}])
+  t.equal(matchesUser(userMap, 'admin', 'admin'), true)
+  t.equal(matchesUser(userMap, 'admin', 'admin2'), false)
+  t.end()
+})
+
 t.test('list', t => {
   const userMap = users2UserMap([{username: 'admin', password: 'admin'}, {username: 'demo', password: 'demo'}])
   t.equal(matchesUser(userMap, 'admin', 'admin'), true)